backend/pyproject.toml
[tool.poetry]
name = "qms-agent-system"
version = "0.1.0"
description = "Medical Device QMS Agent System with ISO 13485 compliance"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.0"

# Google APIs
google-generativeai = "^0.3.2"
google-api-python-client = "^2.115.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"

# LangChain & LangGraph
langchain = "^0.1.0"
langchain-google-genai = "^0.0.6"
langgraph = "^0.0.20"
langchain-community = "^0.0.13"

# Vector DB
chromadb = "^0.4.22"

# Additional utilities
pandas = "^2.1.4"
openpyxl = "^3.1.2"
httpx = "^0.26.0"
redis = "^5.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
